Add more compilation flags and fix errors

+++ src/cpu.h
@@ -1,8 +1,8 @@
 #ifndef __CPU_H__
 #define __CPU_H__
 
-#include <stdint.h>
 #include <stddef.h>
+#include <stdint.h>
 
 /**
  * References:
@@ -37,9 +37,9 @@
 
 #define FLAG_ANY (FLAG_C | FLAG_H | FLAG_N | FLAG_Z)
 
-#define FLAG_IS_SET(flag) (g_cpu.reg.f & flag)
+#define FLAG_IS_SET(flag) (uint8_t)(g_cpu.reg.f & flag)
 #define FLAG_SET(x) (g_cpu.reg.f |= (x))
-#define FLAG_CLEAR(x) (g_cpu.reg.f &= ~(x))
+#define FLAG_CLEAR(x) (g_cpu.reg.f &= (uint8_t)(~(x)))
 
 /**
  * Z80 registers struct.
